refactor: modularize game entities
